refactor: always use `buildSheetsUrl` when creating google sheet urls (#716)
